==> 3.Preprocessing_Modeling_Part1.ipynb <==
<h2> 3. Data Preprocessing </h2>

In this data set, before the modeling phase, I will process the data by separating input and target variables and splitting the test set from the training set. Then, I will standardize the data as I will use certain models (e.g. linear regression) that require standardization, but, first, I'm going to load the data set. 

#Importing required libraries
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import time

#Define a timer
def print_time(start_time, end_time):
    elapsed_time = end_time - start_time
    hours = elapsed_time // 3600
    mins = (elapsed_time - hours*3600)//60
    secs = (elapsed_time - hours*3600 - mins*60) // 1
    
    return print("\nTime elapsed: {} hours {} minutes and {} seconds".format(hours, mins, secs))

from sklearn.model_selection import train_test_split

import warnings
warnings.filterwarnings("ignore")

#Load data set
df = pd.read_csv("saved_files/concrete.csv")
df.head()

#separate inputs from target
X = df.drop('strength', axis=1)
y = df['strength']
#check X and y
print(y.head())
print(X.head())

#split test and train sets
X_train, X_test, y_train, y_test = train_test_split(X, y, train_size = 0.8, random_state = 10)
print("Dimensions of X_train: ", X_train.shape)
print("Dimensions of X_test: ", X_test.shape)
print("Dimensions of y_train: ", y_train.shape)
print("Dimensions of y_test: ", y_test.shape)

#Standardize input variables
X_train_scaled = (X_train - X_train.mean(axis = 0))/X_train.std(axis = 0)
print(X_train_scaled.head())

While I've already standardized the data, I won't directly use the standardized data set since I'll use the *StandardScaler* function in pipelines in the modeling phase.

# <h2> 4. Modeling </h2>

In the modeling phase, I will start with a base model of linear regression. Since this is a regression problem, I will use Root Mean Squared Error (RMSE) as the metric for evaluation with lower RMSEs indicating better models. In all of the models I will use **Cross-Validation** with **10** folds and take the mean RMSE of all 10 results for evaluation purposes. 

#import models & other required packages
from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import Pipeline
from sklearn.metrics import mean_squared_error

from sklearn.linear_model import LinearRegression
from sklearn.neighbors import KNeighborsRegressor
from sklearn.model_selection import GridSearchCV

## <h3> 4.1. Linear Regression </h3>



#define Linear Regression pipe
pipe_linreg = Pipeline([("scaler", StandardScaler()),
                        ("model", LinearRegression())
                        
])

#fit Linear Regression pipe
start_time = time.perf_counter()

pipe_linreg.fit(X_train, y_train)

end_time = time.perf_counter()
print_time(start_time, end_time)


#Predict using fitted Linear Regression model
linreg_preds = pipe_linreg.predict(X_test)

linreg_rmse = mean_squared_error(y_test, linreg_preds, squared = False)
print("Linear Regression RMSE:",linreg_rmse)

#Store results
results = pd.DataFrame({"RMSE" : linreg_rmse}, index = ["Linear Regression"])
results

#for reference, please see the mean compressive strength of concrete
df["strength"].mean()

The base Linear Regression Model gives an RMSE of **10.66**. Now, I will look at other models to improve the RMSE.

## <h3> 4.2. K-Nearest Neighbors (KNN)</h3>

Now, I will use the KNN model with 3 neighbors with uniform weighting. I will define a pipe that will standardize the data before fitting the model.


#define KNN pipe
pipe_knn = Pipeline([("scaler", StandardScaler()),
                         ("model", KNeighborsRegressor(n_neighbors = 3)) #weights = 'uniform' is the default
                        
])

#fit KNN pipe
start_time = time.perf_counter()

pipe_knn.fit(X_train, y_train)

end_time = time.perf_counter()
print_time(start_time, end_time)


#Predict using fitted KNN model
knn_preds = pipe_knn.predict(X_test)

knn_rmse = mean_squared_error(y_test, knn_preds, squared = False)
print("KNN RMSE:",knn_rmse)

#Store results
results.loc["KNN"] = knn_rmse
results

The KNN model gives us an RMSE of **8.56** MPa which is an improvement of over 2 megapascals over the base linear regression model.

## <h3> 4.3. K-Nearest Neighbors (KNN) with GridSearchCV</h3>

Now, I will use the cross validation and tune hyperparameters for the KNN model to see if we have an RMSE improvement.
The hyperparameters I will optimize are the number of neighbors as well as the weighting function.


#Define KNN + GridSearchCV pipe
pipe_knn_grid = Pipeline([("scaler", StandardScaler()),
                     ("model", KNeighborsRegressor())
                        
])
#Define parameter space
params_knn = {'model__n_neighbors': [3, 5, 7, 9, 11, 13, 15, 17, 19],
            'model__weights': ["uniform","distance"]
           }


#use n_jobs = -1 to use all processors
knn_grid = GridSearchCV(pipe_knn_grid, param_grid = params_knn, n_jobs = -1, cv=10, scoring='neg_root_mean_squared_error')

#Fit KNN + GridSearchCV
start_time = time.perf_counter()

knn_grid.fit(X_train, y_train)

end_time = time.perf_counter()
print_time(start_time, end_time)

#Print best parameters
print(knn_grid.best_params_)

#Predict using fitted KNN + GridSearchCV model
knn_grid_preds = knn_grid.best_estimator_.predict(X_test)

knn_grid_rmse = mean_squared_error(y_test, knn_grid_preds, squared = False)
print("KNN + GridSearchCV RMSE:",knn_grid_rmse)

#Store results
results.loc["KNN + GridSearchCV"] = knn_grid_rmse
results

With GridSearchCV, the KNN model substantially improves to an RMSE of **7.38** from the original RMSE of 8.56.

Let's save X_train, X_test, y_train, y_test, ad Results sets so that the second Modeling part can load them directly.

#Save X_train set as a csv file
X_train.to_csv('saved_files/concrete_x_train.csv', index=False)
#Save X_test set as a csv file
X_test.to_csv('saved_files/concrete_x_test.csv', index=False)
#Save y_train set as a csv file
y_train.to_csv('saved_files/concrete_y_train.csv', index=False)
#Save y_test set as a csv file
y_test.to_csv('saved_files/concrete_y_test.csv', index=False)
#Save model results as a csv file
results.to_csv('saved_files/concrete_model_results.csv', index=True)

Next notebook will focus on Random Forest, XGBoost, their optimization and some ensemble model.
