==> 4.Modeling_Part2.ipynb <==
# <h2> 4. Modeling (Part 2) </h2>

First, I'm going to load the data set and then import necessary libraries.

import pandas as pd
import numpy as np
import time

#Define a timer
def print_time(start_time, end_time):
    elapsed_time = end_time - start_time
    hours = elapsed_time // 3600
    mins = (elapsed_time - hours*3600)//60
    secs = (elapsed_time - hours*3600 - mins*60) // 1
    
    return print("\nTime elapsed: {} hours {} minutes and {} seconds".format(hours, mins, secs))

import warnings
warnings.filterwarnings("ignore")

#Load data sets
X_train = pd.read_csv("saved_files/concrete_x_train.csv")
X_test = pd.read_csv("saved_files/concrete_x_test.csv")
y_train = pd.read_csv("saved_files/concrete_y_train.csv")
y_test = pd.read_csv("saved_files/concrete_y_test.csv")
results = pd.read_csv("saved_files/concrete_model_results.csv", index_col = 0)

#Check loaded data sets
print("X_train")
print(X_train.head(), end = "\n\n")
print("X_test")
print(X_test.head(), end = "\n\n")
print("y_train")
print(y_train.head(), end = "\n\n")
print("y_test")
print(y_test.head(), end = "\n\n")
print("Results")
print(results, end = "\n\n")

#import models & other required packages
from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import Pipeline
from sklearn.metrics import mean_squared_error

#Import libraries for models
from sklearn.model_selection import GridSearchCV
from sklearn.ensemble import RandomForestRegressor
#! pip install xgboost #uncomment if xgboost is not already installed 
from xgboost import XGBRegressor

# install and import libraries for hyperparameters tuning
#! pip install scikit-optimize #uncomment if scikit-optimize is not already installed
from skopt import BayesSearchCV
from skopt import space

## <h3> 4.4. Random Forest</h3>

Now, I will use a Random Forest model. Although standardization is not necessarily useful for this model, I will define a pipe that will standardize the data before fitting the model for consistency.


#define Random Forest pipe
pipe_rf = Pipeline([("scaler", StandardScaler()),
                     ("model", RandomForestRegressor(random_state = 10, 
                                                     criterion = "mse"))
                        
])

#fit Random Forest pipe
start_time = time.perf_counter()

pipe_rf.fit(X_train, y_train)

end_time = time.perf_counter()
print_time(start_time, end_time)

#Predict using fitted Random Forest model
rf_preds = pipe_rf.predict(X_test)

rf_rmse = mean_squared_error(y_test, rf_preds, squared = False)
print("Random Forest RMSE:",rf_rmse)

#Store results
results.loc["Random Forest"] = rf_rmse
results

With the Random Forest model, the RMSE is **4.71**, a major improvement over previous models.

## <h3> 4.5. Random Forest with GridSearchCV</h3>

Now, I will use the cross validation and tune hyperparameters for the Random Forest model to see if we have an RMSE improvement.
The hyperparameters I will tune are:


*   Number of trees in the forest (i.e. *n_estimators*)
*   Number of features to consider when looking for the best split (i.e. *max_features*)



#Define Random Forest + GridSearchCV pipe
pipe_rf_grid = Pipeline([("scaler", StandardScaler()),
                          ("model", RandomForestRegressor(random_state = 10, criterion = "mse"))
                        
])
#Define parameter space
params_rf = {'model__n_estimators': range(100, 1001, 50), #from 100 to 1000 in increments of 50
              'model__max_features': [0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9, 1.0]
           }


#use n_jobs = -1 to use all processors
rf_grid = GridSearchCV(pipe_rf_grid, param_grid = params_rf, n_jobs = -1, cv=10, scoring='neg_root_mean_squared_error')

#Fit Random Forest + GridSearchCV
start_time = time.perf_counter()

rf_grid.fit(X_train, y_train)

end_time = time.perf_counter()
print_time(start_time, end_time)

#Print best parameters
print(rf_grid.best_params_)

#Predict using fitted Random Forest + GridSearchCV model
rf_grid_preds = rf_grid.best_estimator_.predict(X_test)

rf_grid_rmse = mean_squared_error(y_test, rf_grid_preds, squared = False)
print("Random Forest + GridSearchCV RMSE:",rf_grid_rmse)

#Store results
results.loc["Random Forest + GridSearchCV"] = rf_grid_rmse
results

#Store predictions for an Ensemble
predictions = pd.DataFrame({"Random Forest + GridSearchCV" : rf_grid_preds})
predictions

With GridSearchCV, the RMSE of the Random Forest model improves to **4.50**, an improvement over the Random Forest model with default parameters.

## <h3> 4.6. Extreme Gradient Boosting (XGBoost) </h3>

Now, I will use an Extreme Gradient Boosting (XGBoost) model. Like before, I will define a pipe that will standardize the data before fitting the model for consistency.


#define XGBoost pipe
pipe_xgb = Pipeline([("scaler", StandardScaler()),
                     ("model", XGBRegressor(random_state = 10))
                        
])

#fit XGBoost pipe
start_time = time.perf_counter()

pipe_xgb.fit(X_train, y_train)

end_time = time.perf_counter()
print_time(start_time, end_time)

#Predict using fitted XGBoost model
xgb_preds = pipe_xgb.predict(X_test)

xgb_rmse = mean_squared_error(y_test, xgb_preds, squared = False)
print("XGBoost RMSE:",xgb_rmse)

#Store results
results.loc["XGBoost"] = xgb_rmse
results

XGBoost with its default hyperparameters gives an RMSE of **4.33**.

## <h3> 4.7. Extreme Gradient Boosting (XGBoost) with BayesSearchCV</h3>

Now, I will optimize some hyperparameters of the Extreme Gradient Boosting (XGBoost) model using BayesSearchCV. BayesSearchCV from scikit-optimize library is used to conduct Bayesian optimization with cross validation.


#Define XGBoost + BayesSearchCV pipe
pipe_xgb_bo = Pipeline([("scaler", StandardScaler()),
                          ("model", XGBRegressor(random_state = 10,
                                                 objective = "reg:squarederror",
                                                 verbosity = 0))
                        
])
#Define parameter space
params_xgb_bo = {'model__n_estimators': space.Integer(100, 1000, 'log-uniform'),
              'model__learning_rate': space.Real(10e-6, 1.0, 'log-uniform'),
              'model__max_depth': space.Integer(0, 50, 'uniform')
           }


#use n_jobs = -1 to use all processors
xgb_bo = BayesSearchCV(pipe_xgb_bo, 
                       params_xgb_bo, 
                       n_iter=50, 
                       cv=10, 
                       random_state = 10, 
                       scoring='neg_root_mean_squared_error')

#Fit XGBoost + BayesSearchCV
start_time = time.perf_counter()

xgb_bo.fit(X_train, y_train)

end_time = time.perf_counter()
print_time(start_time, end_time)

#Print best parameters
print(xgb_bo.best_params_)

#Predict using fitted XGBoost + BayesSearchCV model
xgb_bo_preds = xgb_bo.best_estimator_.predict(X_test)

xgb_bo_rmse = mean_squared_error(y_test, xgb_bo_preds, squared = False)
print("XGBoost + BayesSearchCV RMSE:",xgb_bo_rmse)

#Store results
results.loc["XGBoost + BayesSearchCV"] = xgb_bo_rmse
results

#Store predictions for an Ensemble
predictions["XGBoost + BayesSearchCV"] = xgb_bo_preds
predictions

XGBoost + BayesSearchCV gives an RMSE of **4.14**.

## <h3> 4.8. Ensemble </h3>

Now, I will combine the predictions made by the Random Forest + GridSearchCV model with the ones by XGBoost + BayesSearchCV to improve the results. I will take their mean for the ensemble model.

#Take the average of predictions made by XGBoost and Random Forest + GridSearchCV
ensemble_preds = (predictions.iloc[:, 0] + predictions.iloc[:, 1])/2
predictions["Ensemble"] = ensemble_preds
predictions

ensemble_rmse = mean_squared_error(y_test, ensemble_preds, squared = False)
print("Ensemble", ensemble_rmse)

#Store results
results.loc["Ensemble"] = ensemble_rmse
results

The ensemble model using Random Forest + GridSearchCV and XGBoost + BayesSearchCV gives an RMSE of **4.12**.

Let's save the latest version of the model results and the ensemble predictions as csv files.

#Save model results as a csv file
results.to_csv('saved_files/concrete_model_results.csv', index=True)
#Save ensemble predictions as a csv file
predictions.to_csv('saved_files/concrete_ensemble_predictions.csv', index=False)

# <h2> 5. Conclusion </h2>

As seen the results table above, the Ensemble model of both Random Forest + GridSearchCV and XGBoost + BayesSearchCV obtained the best result with an RMSE of **4.12** while the XGBoost + BayesSearchCV was the second best model with an RMSE of **4.14**.

The Ensemble model improved the RMSE score by around **6.54** MPa (i.e. megapascals) from the base Linear Regression Model.

Next, in another notebook, I will use Neural Networks to try to obtain an even lower RMSE score. 

Overall, this data set was an interesting set.
